Working on Stats about Roll Usage
How many groups each roll is/is not a part of, for now. Right now I
think this mainly is just giving me insight, hopefully for a better way
to actually go from the group/order syntax I have now to actually
implementing a way to show the different groups that make up a legit
order.