Added Sorted Length Array
Added an array that holds the various roll lengths in ascending
order. Could have implemented a form of quick sort on a regular array,
but since I was already creating the array index by index while going
through the file it seemed best to just make it sorted from the start.