Modified Main Loop for Group Only
I'm going to shift the program to do work on the groups alone, not on
the order like I was doing before. The best method I had come up with
for getting the order, and finding groups that made it up, couldn't even
get through the first order. Back to the drawing board.