Starting incrementInt Function
For faster generation of all of the groups, I want to be able to only do
numbers with minRollsForGroup - maxRollsForGroup bits set to 1. I could
technically do this by using the increment array function, with the
array representing the bits that are 1, and then converting the array to
an unsigned int every time, but I want to try and do it this way, since
it is a good challenge and will speed it up in the process.