Completely Moved to Group-Only Main Loop
Main loop before was only doing operations on groups, but it still had
the issue where it was creating a bunch of junk groupings that wouldn't
even have the correct number of rolls needed to make up a group. Even
though it was fairly quick to check and discard numbers that didn't fall
within the correct range, it was still way too many extra numbers.

Now, there is a loop going from minRollsInGroup to maxRollsInGroup. An
array is generated with its indexes = index, and a loop is entered with
that array incrementing up to numberOfRolls. This will give all of the
permutations of a group with X rolls. Each of these is checked, and if
they fit the length criteria they are added to the overall group array.