Added Order Loop
Had a loop that just got valid groups, now have a loop that just grabs
valid orders. It is much much faster than the early early loop that went
through every single number up to the maximum for the number of given
rolls. This takes ~10 seconds to complete.

I want to try doing work on the orders themselves again, I feel like
the faster way to get everything is working backwards from valid order,
to the groups that make it up. Not sure, but this is the start of
testing that.