Checking Time/Memory Needs for Groups in Order
Now that I'm running through all of the orders, I have been doing some
preliminary tests to see how many groups would be valid for various
orders, to get an idea of the time/memory cost if that's a road I want
to go down.

Doesn't look super promising, next update is going to be generating a
linked adjacency list to hold all of the graph info.