More Testing
Didn't go through with the adjacency list, since it appears that it
would take up multiple GBs of space currently. Looking into more methods
for reducing this space.

For now, more testing is being done, with known orders of 17/21/27
rolls, focusing on 27 rolls for now since it is such a massive step up
from the other 2 and if I can figure it out for that, the other 2 should
be fairly trivial.