Made Funcs for Standard Array Creation/Appending
Because of how often I was creating dynamic arrays, and then having to
create seperate variables to keep track of the size/length of the array.
This lead to a lot of repeated code, a lot of extra variables, and a lot
of ways to create little errors.

Next step is to replace all of the current arrays with this new
implemtation and test it.