Fixed realloc Unsafe Behaviour
When resizing arrays in the standard function, I was inputting the
pointer into realloc, and setting that same pointer to the output. This
could cause an issue if realloc-ing fails, since the pointer will now be
NULL, and there is no way to get it back. Now, a temporary pointer is
set to the output, and checked if that is NULL, and therefore can safely
free the old pointer and gracefully exit the program.