Added Delete/FindCompatibleGroups to Trie
Closes #1
For the issue: binary search tree just won't work, however trie looks
promising. It at least finishes the unique pairs problem in a reasonable
time.

Added those two functions to the library, for freeing the whole tree,
and then for finding groups that would work with the given input. For
now it is printing, I will update so that it can do something more
useful.