Added SmallArray
Basically a boolean array, but as compact as can be. The information for
a number is stored within a single bit within a larger array of
uint_32's. All appears to be working, I will used it to offer the
fastest possible check to see if a group already exists.