Implemented intArray into main for Balanced Arrays
One of the downsides of the simple approach (groups that start with X),
was that the earlier rolls had way way larger arrays. To try and combat
this, intArray is now being used to create a balanced array of arrays,
that will always add to the array with the least amount of objects.
Looking at the output, it correctly grabbed/added all of the groups, and
overall they all look fairly balanced. Most are right around 8800, with
a few below in the 6000's, and some above in the 9000's.