Cleaned up Main Group Loop
Old loop used an array of ints to represent a group, and did some work
to use that to find valid groups. Updated it to instead make use of the
nextSetBit functions, which kept the speed about the same, but cleaned
up the loop as a whole quite a bit, and made it more obvious what was
happening from a glance.