Updating Unbiased Check
Before it was going through all of the groupArray to get the first
group, but this left a lot of checks still being duplicated. In the
normal pair, for group at index i, you can automatically discount all of
the groups up to index i - 1 since they would have been checked before.
These new unbiased/biased methods weren't taking advantage of that,
instead always checking every part in every array as long as that array
had an index that wasn't a set bit in the group. Updating this in the
unbiased check instantly shot it up to basically top of the methods,
especially when taking into account how memory effecient/simple it is.