Added fileReaders File
Used for reading in the rolls from the file.
Also, started work on converting all of the single variables for the
order (minGroupSize, minGroupsPerOrder, etc) into one orderStats struct.
Also, added the numberOfRolls and rollList variables to the orderStats
struct, again hoping to cut down on extreneous parameters for some
functions.