Removing Double Biased Array
Took a lot of time to generate, and now that I've been working with the
unbiased array that includes the guard against previous checks and
realized it's already pretty much caught up to that array, it doesn't
seem worth it to continue to pursue that. Especially with how freaking
convoluted that code was.