Testing Sorting RollList by # Groups
Created an initial loop that goes through all of the possible groups,
counts up how many groups each roll is a part of, and then rearranges
the orderStats->rollList to go in descending order of # of groups a roll
is a part of. Doing this method seems to be the absolute best for
speed/cutting down on checks: 80 secs->62 secs, 24.2 billion
checks->12.2 billion checks.

No other sorting is done after the initial roll sorting, it is just the
normal biased array where the earlier rolls will fill their arrays
first.