Created Functions for Re-Ordering rollList
Now that I found that the best method is to sort the rollList itself in
descending order of numGroups a roll is a part of, I optimized and moved
a function out that will collect that information of # of groups, and
then another function for sorting that. All is done in place right
within the orderStats object itself.