- Debug Flag, + withoutRolls, Shrunk GroupArray
Removed the debug flag from the Makefile.
Added a withoutRolls function (plan is to try and work my way backwards
from the order to the groups, see if that works).
When setting group array, I wasn't shrinking it at the end, leaving some
memory just there, also causing an error when trying to use
groupArray->size (which I really shouldn't be doing anyways, it should
be groupArray->length), but the memory should still be cleared.