Updated main recursiveSolve Call
There wasn't a way to get information about the current index/group
being worked on with the old way of just calling recursiveSolve in main.
Updated so that it goes through every group by going through the roll
numbers and then array sizes, and that way I can print out information
to the screen as it runs. It did actually complete, but I didn't get a
time nor a total # because I was printing out too much. This version has
all of that printing commented out, and I am now running it again.