Created int * for ordersWithRoll
Initially did it for just more information, but now I see that when
running for the first index, it gets caught for a really long time at
the limit of orders with roll 0.
Next step is to implement this into the solver, to have it counting the
ordersCountedWithRoll[x] and exit out of index i if ordersCountWithRoll[i] == ordersWithRoll[i]. Actually, just typing that out, probably best to just subtract from the already created ordersWithRoll and check for 0's.