Added ordersWithRoll to Recursive Solve
Because I was able to gather the data for how many orders each roll is a
part of, I am now able to use these values within the recursive solve to
break out of a loop early when there can't be any more orders within the
starting index. For example, with the 29 roll file, index/roll 0 would
get held up a little over 400,000. Using the ordersWithRoll data, I saw
that the hangup lined up with the number of orders involving roll 0.
Seeing this, the recursive function now keeps track of how many orders
are left for each roll, and if it's zero, it will return/skip as much as
possible to do with that roll.