Added Separate Solve Function
Next attempt at simplification/hopefully optimization is to have the
function that is called by main() be one that requires very little. It
will do the work of setting up pointers and arrays and all of that that
the main function really has no use for, so that the parameters are next
to none, and hopefully by splitting it out like that, I can make the
inner recursive solve take a lot more parameters and be ugly, but maybe
using const ints for parameters like numberOfRolls, minGroupLength,
etc. (or maybe just variables global to solver.c?), it could possibly
speed up the solving.