Cleaning up Recursive Solve Function
The scope of a lot of the variables passed to the function was way too
large, they don't need to live beyond the function itself. So, created a
helper function that will take in the needed parameters (orderStats,
groupsWithRoll) and takes care of creating the limited life variables
(alreadyFound, numberFound, etc).