Removed groupArray, groupsWithRoll Returns Pointer
groupArray wasn't being used for anything, so removed it. Instead, that
loop that goes through all of the valid groups is used to create
groupsWithRollBySize. Also, updated that set function to a get function,
since there should be no need for the main function to malloc the memory
for the IntArray**.