Changing threadSolve to Cycle Through Groups
Previous version had all of the work being done with the first thread.
This was because the first parameter being set in the old function was
always a group of 0, index of -1, and numGroups = 0. This meant that
the first thread would just go through the entire thing no problem, and
be done with it. Updated so that it is cycling through all of the groups
within the array at startingIndex.