Working Threads
Code is currently commented out as I compare thread vs. non-thread.
Thread just barely eeks out non-thread, but I think it's beccause I'm
splitting them in a dumb way (1 per groupsWithRoll[i], when in reality
the majority of the time is occuring within groupsWithRoll[0].
Next step is to update the threadSolve function to take in a starting
and ending group index, in addition to startingIndex, that specify the
indexes within groupsWithRoll[i] that should be gone through, rather
than all of them.