Threads to Multiple per startsWithRoll[i]
Before there were just numberOfRolls threads, but the earlier indexes
had way more work to do because of how large they are. So, updated the
loop to now have it deploy threads based on how large the array is.
There is a set number of groups allowed at a time, and for each array it
will deploy its number of groups / number of groups allowed at a time
threads for the different sections of that array. This change makes the
cpu and fans scream, but also cut down the 29 roll test to just over 4
minutes.
TODO: update the theadIds array to include all of these new threads.