Group Split Thread Test
The actual code for group split thread testing is actually commented
out, it turns out that the order that the getSmallArray checks happen
matters. Doing it that way was much quicker, 7 mins -> 4 minutes, but
would miss out on hundreds of orders.

With it only being a 7 minute process, I'm going to go forward with the
slower, but more thorough method.