Fixed File Output for Groups
Was running into issues with all of the work from each thread being
shared (in terms of the group array), which was overwriting and doing
weird things. Solution was to create a temporary array that is passed to
the stack with each new addition, so each is unique.

May be worth looking into whether or not it's worth freeing this memory
after it is done being used, thought for how to do that involves a big
ol' array of all of the pointers that were created, and freeing them all
at the end.