Added Functional Tests Code to Makefile
It will be able to go through the functional tests folder, gather all of
the valid .input files, run them as input in rollcombos and sending
their output to a .check file with the same name. Next steps are to have
the comparisons between the .check and the matching .output, and
preventing the program from spitting out all of those lines with
carriage return if being used this way.